refactor(commands): extract think-back commands into thinkback.rs
