+++ src-rust/crates/commands/src/lib.rs
@@ -242,6 +242,8 @@ mod sandbox;
 pub use sandbox::*;
 mod ultrareview;
 pub use ultrareview::*;
+mod thinkback;
+pub use thinkback::*;
 mod ui_settings;
 use ui_settings::*;
 
@@ -286,8 +288,6 @@ pub struct SecurityReviewCommand;
 pub struct TerminalSetupCommand;
 pub struct ExtraUsageCommand;
 pub struct FastCommand;
-pub struct ThinkBackCommand;
-pub struct ThinkBackPlayCommand;
 pub struct FeedbackCommand;
 pub struct ColorSetCommand;
 // New commands: teleport, btw, ctx-viz, sandbox-toggle
@@ -2585,146 +2585,6 @@ impl SlashCommand for FastCommand {
                 ),
             )
         }
-    }
-}
-
-// ---- /think-back ---------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ThinkBackCommand {
-    fn name(&self) -> &str { "think-back" }
-    fn aliases(&self) -> Vec<&str> { vec!["thinkback"] }
-    fn description(&self) -> &str { "Show thinking traces from previous responses in this session" }
-    fn help(&self) -> &str {
-        "Usage: /think-back [n]\n\n\
-         Displays the thinking/reasoning traces from the most recent model responses.\n\
-         Pass a number to show the Nth most recent thinking block."
-    }
-
-    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let n: usize = args.trim().parse().unwrap_or(1).max(1);
-
-        // Scan messages for thinking blocks
-        let thinking_blocks: Vec<(usize, String)> = ctx
-            .messages
-            .iter()
-            .enumerate()
-            .filter(|(_, m)| m.role == claurst_core::types::Role::Assistant)
-            .filter_map(|(idx, m)| {
-                let blocks = m.get_thinking_blocks();
-                if blocks.is_empty() {
-                    return None;
-                }
-                let thinking: String = blocks
-                    .iter()
-                    .filter_map(|b| {
-                        if let claurst_core::types::ContentBlock::Thinking { thinking, .. } = b {
-                            Some(thinking.as_str())
-                        } else {
-                            None
-                        }
-                    })
-                    .collect::<Vec<_>>()
-                    .join("\n\n");
-                if thinking.is_empty() { None } else { Some((idx, thinking)) }
-            })
-            .collect();
-
-        if thinking_blocks.is_empty() {
-            return CommandResult::Message(
-                "No thinking traces found in this session.\n\
-                 Thinking traces appear when the model uses extended thinking mode.\n\
-                 Try asking Claurst to 'think step by step' or 'think carefully'."
-                    .to_string(),
-            );
-        }
-
-        // Show the Nth most recent (1-indexed)
-        let total = thinking_blocks.len();
-        let target_idx = total.saturating_sub(n);
-        let (msg_idx, trace) = &thinking_blocks[target_idx];
-
-        CommandResult::Message(format!(
-            "Thinking trace ({n} of {total} found, from message {msg}):\n\
-             ─────────────────────────────────────\n\
-             {trace}\n\
-             ─────────────────────────────────────\n\
-             Use /think-back <n> to see older traces.",
-            n = n,
-            total = total,
-            msg = msg_idx + 1,
-            trace = trace,
-        ))
-    }
-}
-
-// ---- /thinkback-play -----------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ThinkBackPlayCommand {
-    fn name(&self) -> &str { "thinkback-play" }
-    fn description(&self) -> &str { "Replay a thinking trace as an animated walkthrough" }
-    fn help(&self) -> &str {
-        "Usage: /thinkback-play [n]\n\n\
-         Replays a previous thinking trace, formatted for easy reading.\n\
-         Pass a number to replay the Nth most recent trace."
-    }
-
-    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let n: usize = args.trim().parse().unwrap_or(1).max(1);
-
-        let thinking_blocks: Vec<String> = ctx
-            .messages
-            .iter()
-            .filter(|m| m.role == claurst_core::types::Role::Assistant)
-            .filter_map(|m| {
-                let blocks = m.get_thinking_blocks();
-                if blocks.is_empty() {
-                    return None;
-                }
-                let t: String = blocks
-                    .iter()
-                    .filter_map(|b| {
-                        if let claurst_core::types::ContentBlock::Thinking { thinking, .. } = b {
-                            Some(thinking.as_str())
-                        } else {
-                            None
-                        }
-                    })
-                    .collect::<Vec<_>>()
-                    .join("\n\n");
-                if t.is_empty() { None } else { Some(t) }
-            })
-            .collect();
-
-        if thinking_blocks.is_empty() {
-            return CommandResult::Message(
-                "No thinking traces to replay in this session.".to_string(),
-            );
-        }
-
-        let total = thinking_blocks.len();
-        let idx = total.saturating_sub(n);
-        let trace = &thinking_blocks[idx];
-
-        // Format the trace with step numbering
-        let steps: Vec<&str> = trace.split('\n').filter(|l| !l.trim().is_empty()).collect();
-        let mut formatted = format!(
-            "Thinking Trace Replay ({}/{total})\n\
-             ══════════════════════════════════\n",
-            n,
-            total = total
-        );
-        for (i, step) in steps.iter().enumerate() {
-            formatted.push_str(&format!("  Step {}: {}\n", i + 1, step));
-        }
-        formatted.push_str("══════════════════════════════════\n");
-        formatted.push_str(&format!(
-            "{} steps shown. Use /think-back for raw traces.",
-            steps.len()
-        ));
-
-        CommandResult::Message(formatted)
     }
 }
